+++ parsers/frontmatter.h
@@ -1,0 +1,19 @@
+/*
+ *   Copyright (c) 2022, Masatake YAMATO <yamato@redhat.com>
+ *
+ *   This source code is released for free distribution under the terms of the
+ *   GNU General Public License version 2 or (at your option) any later version.
+ *
+ *   Frontmatter parser interface exported to the other parsers
+ */
+
+#ifndef CTAGS_FRONTMATTER_H
+#define CTAGS_FRONTMATTER_H
+
+#include "general.h"
+
+typedef enum {
+	FRONTMATTER_TITLE_KIND,
+} frontmatterKind;
+
+#endif